todo_backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
